Move structs using Windows API to api module
