+++ src/window_picker/tests/mod.rs
@@ -1,2 +1,1 @@
-mod native_ui_tests;
 mod window_picker_tests;
